gate Darwin-dependent API behind conditionals
